chore(core): move enum classes
