+++ app/src/main/java/com/osfans/trime/ime/core/Trime.java
@@ -61,8 +61,8 @@ import com.osfans.trime.KeyboardView;
 import com.osfans.trime.R;
 import com.osfans.trime.Rime;
 import com.osfans.trime.Speech;
-import com.osfans.trime.enums.InlineModeType;
-import com.osfans.trime.enums.WindowsPositionType;
+import com.osfans.trime.ime.enums.InlineModeType;
+import com.osfans.trime.ime.enums.WindowsPositionType;
 import com.osfans.trime.settings.components.ColorPickerDialog;
 import com.osfans.trime.settings.components.SchemaPickerDialog;
 import com.osfans.trime.settings.components.ThemePickerDialog;
